requirements.txt
Flask
cryptography
psycopg2-binary
python-dotenv